+++ src/main/java/net/silentchaos512/gear/block/compounder/CompounderBlock.java
@@ -19,7 +19,6 @@ import net.minecraft.util.*;
 import net.minecraft.util.math.BlockPos;
 import net.minecraft.util.math.BlockRayTraceResult;
 import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextFormatting;
 import net.minecraft.world.IBlockReader;
 import net.minecraft.world.World;
 import net.minecraftforge.fml.network.NetworkHooks;
@@ -80,7 +79,7 @@ public class CompounderBlock extends Block {
 
     @Override
     public void addInformation(ItemStack stack, @Nullable IBlockReader worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
-        tooltip.add(TextUtil.withColor(TextUtil.misc("wip"), TextFormatting.RED));
+        TextUtil.addWipText(tooltip);
 
         Set<String> catNameSet = this.info.getCategories().stream().map(IMaterialCategory::getName).collect(Collectors.toSet());
         String catStr = Strings.join(catNameSet, ", ");