+++ src/main/java/net/silentchaos512/gear/item/CompoundMaterialItem.java
@@ -9,7 +9,6 @@ import net.minecraft.nbt.ListNBT;
 import net.minecraft.util.NonNullList;
 import net.minecraft.util.ResourceLocation;
 import net.minecraft.util.text.ITextComponent;
-import net.minecraft.util.text.TextFormatting;
 import net.minecraft.util.text.TranslationTextComponent;
 import net.minecraft.world.World;
 import net.minecraftforge.common.util.Constants;
@@ -127,7 +126,7 @@ public class CompoundMaterialItem extends Item implements IColoredMaterialItem {
 
     @Override
     public void addInformation(ItemStack stack, @Nullable World worldIn, List<ITextComponent> tooltip, ITooltipFlag flagIn) {
-        tooltip.add(TextUtil.withColor(TextUtil.misc("wip"), TextFormatting.RED));
+        TextUtil.addWipText(tooltip);
 
         Collection<IMaterialInstance> materials = getSubMaterials(stack);