+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+### Added
+- Config to disable WIP tooltip on some items
+
 ## [2.6.21] - 2021-04-21
 ### Fixed
 - Conditional traits sometimes not appearing gear items (magmatic, soft, etc.) [#372, #359]